block: Switch to noop I/O scheduler when the screen turns off
This driver changes the I/O scheduler for a specified request queue to
Noop when the screen turns off in order to improve power consumption while
the screen is off, as well as reduce resume latency when the screen is
turned back on. After the screen is turned back on, the driver will change
the I/O scheduler back to what it originally was after a 10000ms delay.

Change-Id: I73a76e8dcac865ab730d825f98118376e3aa140e
Signed-off-by: Sultanxda <sultanxda@gmail.com>
Signed-off-by: Ícaro Hoff <icarohoff@gmail.com>